Добавлены репозитории для моделей

+++ src/main/java/com/example/embedika_test/dao/model/CarModel.java
@@ -17,7 +17,7 @@ public class CarModel {
     @Id
     @GeneratedValue
     @Column(name = "model_id")
-    private Long modelId;
+    private Integer modelId;
     @Column(name = "model_name")
     private String name;
 }
